entities: Update Spy arguments storing mechanism
The current approach of storing call arguments directly in `Spy` is
not flexible enough.
This commit adds `MethodCall` entity that stores call arguments.
Now `Spy` stores an array of method calls. That gives us the ability
to count calls out-of-box and now `Spy` stores all calls instead of the
last one.